AG-9343 - Clean up bar gallery example names
