--- packages/ag-charts-website/src/content/gallery/_examples/column-with-negative-values/data.ts
@@ -1,10 +1,0 @@
-export function getData(): any[] {
-    return [
-        { month: 'Jul', men: 79011, women: 3808, menDelta: 149, womenDelta: 62 },
-        { month: 'Aug', men: 79166, women: 3841, menDelta: 155, womenDelta: 33 },
-        { month: 'Sep', men: 79514, women: 3813, menDelta: 348, womenDelta: -28 },
-        { month: 'Oct', men: 79945, women: 3856, menDelta: 431, womenDelta: 43 },
-        { month: 'Nov', men: 79869, women: 3796, menDelta: -76, womenDelta: -60 },
-        { month: 'Dec', men: 79701, women: 3783, menDelta: -168, womenDelta: -13 },
-    ];
-}
