+++ packages/ag-charts-website/src/content/gallery/_examples/grouped-horizontal-bar/data.ts
@@ -1,0 +1,9 @@
+export function getData() {
+    return [
+        { type: 'Whole Economy', total: 2.4, regular: 3.2 },
+        { type: 'Public Sector', total: 3.3, regular: 3.4 },
+        { type: 'Construction', total: 1.8, regular: 2.6 },
+        { type: 'Private Sector', total: 2.7, regular: 3.5 },
+        { type: 'Manufacturing', total: 2.7, regular: 2.9 },
+    ];
+}
