+++ packages/ag-charts-website/src/content/gallery/_examples/grouped-horizontal-bar/index.html
