--- packages/ag-charts-website/src/content/gallery/_examples/simple-column/data.ts
@@ -1,12 +1,0 @@
-export function getData() {
-    return [
-        { year: '2016', visitors: 46636720 },
-        { year: '2017', visitors: 48772922 },
-        { year: '2018', visitors: 50800193 },
-        { year: '2019', visitors: 48023342 },
-        { year: '2020', visitors: 47271912 },
-        { year: '2021', visitors: 47155093 },
-        { year: '2022', visitors: 49441678 },
-        { year: '2023', visitors: 50368190 },
-    ];
-}
