--- packages/ag-charts-website/src/content/gallery/_examples/simple-column/main.ts
@@ -1,54 +1,0 @@
-import { AgEnterpriseCharts, AgChartOptions } from 'ag-charts-enterprise';
-import { getData } from './data';
-
-function formatNumber(value: number) {
-    value /= 1000_000;
-    return `${Math.floor(value)}M`;
-}
-
-const options: AgChartOptions = {
-    container: document.getElementById('myChart'),
-    data: getData(),
-    title: {
-        text: 'Total Visitors to Museums and Galleries',
-    },
-    footnote: {
-        text: 'Source: Department for Digital, Culture, Media & Sport',
-    },
-    series: [
-        {
-            type: 'bar',
-            xKey: 'year',
-            yKey: 'visitors',
-            label: {
-                formatter: ({ value }) => formatNumber(value),
-            },
-            tooltip: {
-                renderer: ({ yValue, xValue }) => {
-                    return { title: xValue, content: formatNumber(yValue) };
-                },
-            },
-        },
-    ],
-    axes: [
-        {
-            type: 'category',
-            position: 'bottom',
-            title: {
-                text: 'Year',
-            },
-        },
-        {
-            type: 'number',
-            position: 'left',
-            title: {
-                text: 'Total Visitors',
-            },
-            label: {
-                formatter: ({ value }) => formatNumber(value),
-            },
-        },
-    ],
-};
-
-AgEnterpriseCharts.create(options);
