+++ packages/ag-charts-website/src/content/gallery/_examples/simple-horizontal-bar/data.ts
@@ -1,0 +1,11 @@
+export function getData() {
+    return [
+        { type: 'Manager', earnings: 954 },
+        { type: 'Engineer', earnings: 844 },
+        { type: 'Entrepreneur', earnings: 699 },
+        { type: 'Lawyer', earnings: 503 },
+        { type: 'Teacher', earnings: 501 },
+        { type: 'Freelancer', earnings: 407 },
+        { type: 'Doctor', earnings: 358 },
+    ];
+}
