+++ packages/ag-charts-website/src/content/gallery/_examples/simple-horizontal-bar/index.html
