+++ packages/ag-charts-website/src/content/gallery/_examples/stacked-bar/data.ts
@@ -1,28 +1,51 @@
 export function getData() {
     return [
         {
-            type: 'Terraced House',
-            ownerOccupied: 1013,
-            privateRented: 822,
-            localAuthority: 168,
-            housingAssociation: 295,
-        },
-        { type: 'Detached House', ownerOccupied: 3655, privateRented: 294, localAuthority: 4, housingAssociation: 19 },
-        { type: 'Bungalow', ownerOccupied: 1486, privateRented: 214, localAuthority: 180, housingAssociation: 228 },
-        { type: 'Converted Flat', ownerOccupied: 252, privateRented: 539, localAuthority: 37, housingAssociation: 100 },
-        {
-            type: 'Low Rise Flat',
-            ownerOccupied: 905,
-            privateRented: 1139,
-            localAuthority: 621,
-            housingAssociation: 919,
+            station: 'Finsbury\nPark',
+            early: 1454,
+            morningPeak: 16644,
+            interPeak: 9338,
+            afternoonPeak: 6346,
+            evening: 3547,
         },
         {
-            type: 'High Rise Flat',
-            ownerOccupied: 97,
-            privateRented: 134,
-            localAuthority: 109,
-            housingAssociation: 68,
+            station: 'Seven\nSisters',
+            early: 1927,
+            morningPeak: 7581,
+            interPeak: 5421,
+            afternoonPeak: 3245,
+            evening: 2036,
+        },
+        {
+            station: 'Tottenham\nHale',
+            early: 836,
+            morningPeak: 12740,
+            interPeak: 14964,
+            afternoonPeak: 12790,
+            evening: 8428,
+        },        {
+            station: 'Warren\nStreet',
+            early: 108,
+            morningPeak: 1210,
+            interPeak: 5902,
+            afternoonPeak: 10947,
+            evening: 5574,
+        },
+        {
+            station: 'Oxford\nCircus',
+            early: 170,
+            morningPeak: 1796,
+            interPeak: 26616,
+            afternoonPeak: 32269,
+            evening: 30665,
+        },
+        {
+            station: 'Green\nPark',
+            early: 252,
+            morningPeak: 1911,
+            interPeak: 11971,
+            afternoonPeak: 19749,
+            evening: 11582,
         },
     ];
 }
