--- packages/ag-charts-website/src/content/gallery/_examples/stacked-column/data.ts
@@ -1,51 +1,0 @@
-export function getData() {
-    return [
-        {
-            station: 'Finsbury \nPark',
-            early: 1454,
-            morningPeak: 16644,
-            interPeak: 9338,
-            afternoonPeak: 6346,
-            evening: 3547,
-        },
-        {
-            station: 'Seven \nSisters',
-            early: 1927,
-            morningPeak: 7581,
-            interPeak: 5421,
-            afternoonPeak: 3245,
-            evening: 2036,
-        },
-        {
-            station: 'Tottenham \nHale',
-            early: 836,
-            morningPeak: 12740,
-            interPeak: 14964,
-            afternoonPeak: 12790,
-            evening: 8428,
-        },        {
-            station: 'Warren \nStreet',
-            early: 108,
-            morningPeak: 1210,
-            interPeak: 5902,
-            afternoonPeak: 10947,
-            evening: 5574,
-        },
-        {
-            station: 'Oxford \nCircus',
-            early: 170,
-            morningPeak: 1796,
-            interPeak: 26616,
-            afternoonPeak: 32269,
-            evening: 30665,
-        },
-        {
-            station: 'Green \nPark',
-            early: 252,
-            morningPeak: 1911,
-            interPeak: 11971,
-            afternoonPeak: 19749,
-            evening: 11582,
-        },
-    ];
-}
