+++ packages/ag-charts-website/src/content/gallery/_examples/stacked-horizontal-bar/data.ts
@@ -1,0 +1,28 @@
+export function getData() {
+    return [
+        {
+            type: 'Terraced House',
+            ownerOccupied: 1013,
+            privateRented: 822,
+            localAuthority: 168,
+            housingAssociation: 295,
+        },
+        { type: 'Detached House', ownerOccupied: 3655, privateRented: 294, localAuthority: 4, housingAssociation: 19 },
+        { type: 'Bungalow', ownerOccupied: 1486, privateRented: 214, localAuthority: 180, housingAssociation: 228 },
+        { type: 'Converted Flat', ownerOccupied: 252, privateRented: 539, localAuthority: 37, housingAssociation: 100 },
+        {
+            type: 'Low Rise Flat',
+            ownerOccupied: 905,
+            privateRented: 1139,
+            localAuthority: 621,
+            housingAssociation: 919,
+        },
+        {
+            type: 'High Rise Flat',
+            ownerOccupied: 97,
+            privateRented: 134,
+            localAuthority: 109,
+            housingAssociation: 68,
+        },
+    ];
+}
