+++ packages/ag-charts-website/src/content/gallery/_examples/stacked-horizontal-bar/index.html
